# api.yaml
openapi: 3.0.3
info:
  title: Art Explorer API
  description: API para la aplicación de exploración de arte, gestión de colecciones de usuarios y catálogo de obras.
  version: 1.0.0
servers:
  - url: https://api.artexplorer.com
    description: Servidor de Producción
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    webhookSecret:
      type: apiKey
      in: header
      name: x-webhook-secret
      description: Secret key for validating Supabase webhook requests
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: '550e8400-e29b-41d4-a716-446655440000'
    CategoryObject:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
...
    Artwork:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        title:
          type: string
        author:
          type: string
        year:
          type: string
        country:
          type: string
        period:
          type: string
        technique:
          type: string
        dimensions:
          type: string
        imageUrl:
          type: string
          format: uri
        description:
          type: string
        category:
          $ref: '#/components/schemas/CategoryObject'
        subcategory:
          $ref: '#/components/schemas/CategoryObject'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # --- User Collection ---
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Invalid image format'
        message:
          type: string
          example: 'The uploaded file must be a valid JPG, PNG, or HEIC image'
    RecognizeArtworkRequest:
      type: object
      required:
        - image
        - localUri
      properties:
        image:
          type: string
          format: binary
          description: 'Imagen capturada del monumento u obra de arte (JPG, PNG, HEIC - máximo 10MB)'
        localUri:
          type: string
          description: 'URI local del dispositivo donde se guardó la imagen (file:// o content://)'
          example: 'file:///storage/emulated/0/DCIM/Camera/IMG_20241209_153045.jpg'
    RecognizedArtworkData:
      type: object
      description: 'Datos de la obra identificada - null si no se pudo identificar'
      nullable: true
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        title:
          type: string
          example: 'Las Meninas'
        artist:
          type: string
          example: 'Diego Velázquez'
        year:
          type: string
          example: '1656'
        period:
          type: string
          example: 'Spanish Baroque'
          description: 'Período artístico o movimiento'
        description:
          type: string
          example: 'One of the masterpieces of the Spanish Golden Age. It shows Infanta Margaret surrounded by her ladies-in-waiting (meninas) and other members of the court, including the painter himself. The complex composition with multiple viewpoints, the masterful use of light, and the mirror reflecting King Philip IV and Queen Mariana of Austria, challenge the perception of space and reality.'
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.98
          description: 'Nivel de confianza del reconocimiento (0.0 - 1.0)'
        tags:
          type: array
          items:
            type: string
          example:
            [
              'Spanish Baroque',
              'Court Portrait',
              'Oil on Canvas',
              '17th Century',
              'Prado Museum',
            ]
          description: 'Categorías, estilos y metadata adicional'
        capturedImageUrl:
          type: string
          example: 'file:///storage/emulated/0/DCIM/Camera/IMG_20241209_153045.jpg'
          description: 'URI local del dispositivo donde se encuentra la imagen capturada'
        identifiedAt:
          type: string
          format: date-time
          example: '2025-11-30T14:30:00Z'
    RecognizedArtworkResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
          description: 'Indica si la operación fue exitosa a nivel técnico'
        identified:
          type: boolean
          example: true
          description: 'Indica si la obra fue identificada por la IA (true) o no se pudo reconocer (false)'
        artwork:
          $ref: '#/components/schemas/RecognizedArtworkData'
        savedToCollection:
          type: boolean
          example: true
          description: 'Indica si la obra se guardó en la colección del usuario (solo true si identified=true)'
        message:
          type: string
          example: 'Artwork successfully identified and saved to collection'
          description: 'Mensaje descriptivo del resultado para el usuario'

    # --- Users ---
...
paths:
...
  /api/camera/recognize:
    post:
      tags:
        - Camera Recognition
      summary: Reconocer monumento u obra de arte desde imagen capturada
      description: |
        Recibe una imagen capturada por el usuario, la procesa mediante IA para identificar
        el monumento u obra de arte, y automáticamente la guarda en la colección del usuario
        si es reconocida exitosamente.

        **Flujo:**
        1. Usuario captura foto desde la cámara y la guarda localmente en el dispositivo
        2. Frontend envía imagen (para análisis) y localUri (para almacenamiento) como multipart/form-data
        3. Backend procesa imagen con IA y busca coincidencias
        4. Si identifica: guarda localUri en colección y retorna datos completos
        5. Si no identifica: retorna identified=false con mensaje descriptivo

        **Limitaciones:**
        - Tamaño máximo: 10MB
        - Formatos soportados: JPG, PNG, HEIC
        - Rate limiting aplicado según tier de suscripción del usuario
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: language
          schema:
            type: string
            pattern: '^[a-z]{2}$'
            default: 'es'
          description: Código ISO 639-1 del idioma (es, en, fr) para los resultados
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RecognizeArtworkRequest'
            encoding:
              image:
                contentType: image/jpeg, image/png, image/heic
      responses:
        '200':
          description: |
            Operación exitosa - puede indicar obra identificada (identified=true)
            o no identificada (identified=false)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecognizedArtworkResponse'
              examples:
                identified:
                  summary: Obra identificada exitosamente
                  value:
                    success: true
                    identified: true
                    artwork:
                      id: '550e8400-e29b-41d4-a716-446655440000'
                      title: 'Las Meninas'
                      artist: 'Diego Velázquez'
                      year: '1656'
                      period: 'Spanish Baroque'
                      description: 'One of the masterpieces of the Spanish Golden Age. It shows Infanta Margaret surrounded by her ladies-in-waiting (meninas) and other members of the court, including the painter himself. The complex composition with multiple viewpoints, the masterful use of light, and the mirror reflecting King Philip IV and Queen Mariana of Austria, challenge the perception of space and reality.'
                      confidence: 0.98
                      tags:
                        [
                          'Spanish Baroque',
                          'Court Portrait',
                          'Oil on Canvas',
                          '17th Century',
                          'Prado Museum',
                        ]
                      capturedImageUrl: 'https://storage.artexplorer.com/captures/user123/abc123.jpg'
                      identifiedAt: '2025-11-30T14:30:00Z'
                    savedToCollection: true
                    message: 'Artwork successfully identified and saved to collection'
                not_identified:
                  summary: Obra no pudo ser identificada
                  value:
                    success: true
                    identified: false
                    artwork: null
                    savedToCollection: false
                    message: 'Artwork could not be identified. Try with better lighting or a different angle.'
        '400':
          description: Solicitud inválida - imagen corrupta, formato no soportado, etc.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_format:
                  summary: Formato de imagen no válido
                  value:
                    success: false
                    error: 'invalid_image_format'
                    message: 'The uploaded file must be a valid JPG, PNG, or HEIC image'
                missing_image:
                  summary: Falta el campo image
                  value:
                    success: false
                    error: 'missing_required_field'
                    message: "The 'image' field is required"
        '401':
          description: No autenticado - token JWT inválido o faltante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'unauthorized'
                message: 'Valid authentication token required'
        '413':
          description: Payload demasiado grande - imagen excede 10MB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'payload_too_large'
                message: 'Image file size must not exceed 10MB'
        '429':
          description: Demasiadas solicitudes - rate limit excedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'rate_limit_exceeded'
                message: 'Recognition limit reached. Upgrade to premium for unlimited scans.'
        '500':
          description: Error interno del servidor - fallo en procesamiento IA o base de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'internal_server_error'
                message: 'An unexpected error occurred while processing the image. Please try again.'
  /api/camera/artwork/details:
    post:
      tags:
        - Camera Recognition
      summary: Obtener información detallada de una obra identificada
      description: |
        Recibe información básica de una obra (título, autor, descripción) y retorna
        información enriquecida adicional como contexto histórico, técnica, ubicación,
        obras relacionadas, y datos curiosos.

        **Uso típico:**
        - Usuario ve la información inicial de una obra identificada
        - Usuario solicita "más información" sobre la obra
        - Frontend envía los datos básicos disponibles
        - Backend enriquece con contexto adicional mediante IA y bases de datos

        **Limitaciones:**
        - Rate limiting aplicado según tier de suscripción
        - Requiere al menos título o autor para funcionar
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: language
          schema:
            type: string
            pattern: '^[a-z]{2}$'
            default: 'es'
          description: Código ISO 639-1 del idioma (es, en, fr) para los resultados enriquecidos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: Título de la obra
                  example: 'Las Meninas'
                artist:
                  type: string
                  description: Nombre del artista/creador
                  example: 'Diego Velázquez'
                description:
                  type: string
                  description: Descripción inicial de la obra
                  example: 'One of the masterpieces of the Spanish Golden Age...'
                year:
                  type: string
                  description: Año de creación (opcional)
                  example: '1656'
                artworkId:
                  type: string
                  description: ID de la obra en la base de datos (si ya fue identificada previamente)
                  example: '550e8400-e29b-41d4-a716-446655440000'
      responses:
        '200':
          description: Información detallada obtenida exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  detailedInfo:
                    type: object
                    properties:
                      historicalContext:
                        type: string
                        description: Contexto histórico y cultural de la época
                        example: 'Created during the reign of Philip IV of Spain, during the Spanish Golden Age...'
                      visualDescription:
                        type: string
                        description: Descripción visual detallada de la obra (composición, colores, elementos, figuras)
                        example: 'The painting depicts a large room in the Royal Alcazar of Madrid. In the center, the Infanta Margarita Teresa is surrounded by her ladies-in-waiting...'
                      symbolism:
                        type: string
                        description: Simbolismo e interpretación de la obra
                        example: 'The mirror in the background reflects the royal couple, suggesting multiple perspectives...'
                      interestingFacts:
                        type: array
                        items:
                          type: string
                        example:
                          - 'Picasso created 58 interpretations of Las Meninas in 1957'
                          - 'The painting includes a self-portrait of Velázquez himself'
                          - 'It was originally titled "The Family of Philip IV"'
                      artistBio:
                        type: string
                        description: Breve biografía del artista
                        example: 'Diego Rodríguez de Silva y Velázquez (1599-1660) was a Spanish painter...'
        '400':
          description: Solicitud inválida - faltan campos requeridos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'missing_required_field'
                message: 'Title is required to fetch artwork details'
        '401':
          description: No autenticado - token JWT inválido o faltante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'unauthorized'
                message: 'Valid authentication token required'
        '404':
          description: Obra no encontrada en la base de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'artwork_not_found'
                message: 'Could not find detailed information for the specified artwork'
        '429':
          description: Demasiadas solicitudes - rate limit excedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'rate_limit_exceeded'
                message: 'Detail request limit reached. Please try again later.'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'internal_server_error'
                message: 'An unexpected error occurred while fetching artwork details'

  # ----------------------------
  # EXPLORE SCREEN
  # ----------------------------
...
  /api/artworks/{id}:
    get:
      tags:
        - Artworks Catalog
      summary: Obtener detalle de una obra específica
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - in: query
          name: language
          schema:
            type: string
            default: 'es'
      responses:
        '200':
          description: Detalle de la obra
          content:
            application/json:
              schema:
                type: object
                properties:
                  artwork:
                    $ref: '#/components/schemas/Artwork'
  /api/artworks/search:
    get:
      tags:
        - Artworks Catalog
      summary: Buscar obras por título o autor
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: Query de búsqueda
        - in: query
          name: language
          schema:
            type: string
            default: 'es'
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Resultados de búsqueda
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Artwork'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      limit:
                        type: integer
                      offset:
                        type: integer
                      hasMore:
                        type: boolean
                  query:
                    type: string
  /api/artworks/daily-recommendation:
    get:
      tags:
        - Artworks Catalog
      summary: Obtener recomendación diaria de obra de arte
      description: |
        Retorna una obra de arte seleccionada aleatoriamente como recomendación del día.

        **Características:**
        - La recomendación se genera automáticamente y es la misma para todos los usuarios durante el día
        - Se renueva automáticamente cada día
        - Requiere autenticación JWT
        - Soporta múltiples idiomas para las traducciones

        **Cache:**
        - La recomendación se cachea en memoria para optimizar rendimiento
        - Cache válido hasta que cambie la fecha (UTC)
        - Si el servidor se reinicia, se genera una nueva recomendación automáticamente
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: language
          schema:
            type: string
            default: 'es'
          description: Código ISO 639-1 del idioma (es, en, fr) para obtener las traducciones
      responses:
        '200':
          description: Recomendación diaria obtenida exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  artwork:
                    $ref: '#/components/schemas/Artwork'
              example:
                artwork:
                  id: '550e8400-e29b-41d4-a716-446655440000'
                  title: 'La noche estrellada'
                  author: 'Vincent van Gogh'
                  year: '1889'
                  country: 'Francia'
                  period: 'Postimpresionismo'
                  technique: 'Óleo sobre lienzo'
                  dimensions: '73.7 cm × 92.1 cm'
                  imageUrl: 'https://example.com/starry-night.jpg'
                  description: 'Una de las pinturas más famosas del arte moderno...'
                  category:
                    id: 'pintura'
                    name: 'Pintura'
                    icon: '🎨'
                  createdAt: '2025-01-01T00:00:00Z'
                  updatedAt: '2025-01-01T00:00:00Z'
        '401':
          description: No autenticado - JWT inválido o no proporcionado
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: 'Unauthorized'

  # ----------------------------
  # USER COLLECTION
  # ----------------------------
...